+++ src/core/analysis/Token.cpp
@@ -111,7 +111,7 @@ namespace Lucene
     
     void Token::setTermBuffer(const String& buffer)
     {
-        int32_t length = buffer.size();
+        int32_t length = (int32_t)buffer.size();
         growTermBuffer(length);
         MiscUtils::arrayCopy(buffer.begin(), 0, _termBuffer.get(), 0, length);
         _termLength = length;
